+++ src/main/java/com/ugcs/gprvisualizer/app/commands/EdgeFinder.java
@@ -23,17 +23,17 @@ public class EdgeFinder implements Command {
 		for (int i = 0; i < traces.size(); i++) {
 			Trace trace = traces.get(i);
 			float[] values = trace.getNormValues();
-			trace.edge = new int[values.length];
+			trace.edge = new byte[values.length];
 			
 			int mxind = 0;
 			for (int s = 1; s < values.length; s++) {
 				
-				int s1 = (int) Math.signum(values[s - 1]);
-				int s2 = (int) Math.signum(values[s]);
+				byte s1 = (byte) Math.signum(values[s - 1]);
+				byte s2 = (byte) Math.signum(values[s]);
 				
 				if (s1 != s2) {
-					trace.edge[s] = s1 > s2 ? 1 : 2;
-					trace.edge[mxind] = (values[mxind]) < 0 ? 3 : 4;
+					trace.edge[s] =  s1 > s2 ? (byte) 1 : 2;
+					trace.edge[mxind] = (values[mxind]) < 0 ? (byte) 3 : 4;
 					mxind = s;
 				}
